+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Rename `layout.zip` to `presentationML.zip`.
+- Update electron doc: you should set `webSecurity: false`.
+
 ## 0.1.1
 
 - Stop trying if not found current slides' `uuid`.